+++ include/pbcopper/parallel/FireAndForget.h
@@ -1,5 +1,3 @@
-// Author: Lance Hepler & Armin Töpfer
-
 #ifndef PBCOPPER_PARALLEL_FIREANDFORGET_H
 #define PBCOPPER_PARALLEL_FIREANDFORGET_H
 
@@ -8,7 +6,9 @@
 #include <atomic>
 #include <condition_variable>
 #include <cstddef>
+#include <cstdint>
 #include <exception>
+#include <functional>
 #include <future>
 #include <mutex>
 #include <queue>
@@ -25,7 +25,7 @@ private:
 
 public:
     FireAndForget(const size_t size, const size_t mul = 2)
-        : exc{nullptr}, sz{size * mul}, abort{false}, thrown{false}
+        : exc{nullptr}, numThreads{size}, sz{size * mul}, abort{false}, thrown{false}
     {
         for (size_t i = 0; i < size; ++i) {
             threads.emplace_back(std::thread([this]() {
@@ -107,6 +107,8 @@ public:
         }
     }
 
+    size_t NumThreads() const { return numThreads; }
+
 private:
     TTask PopTask()
     {
@@ -135,10 +137,22 @@ private:
     std::condition_variable pushed;
     std::exception_ptr exc;
     std::mutex m;
+    size_t numThreads;
     size_t sz;
     std::atomic_bool abort;
     std::atomic_bool thrown;
 };
+
+///
+/// \brief Use an existing FireAndForget to dispatch [0, numEntries) callbacks.
+///        Returns after all dispatched jobs finished.
+///
+/// \param faf         reference to FaF, nullptr allowed
+/// \param numEntries  number of submissions
+/// \param callback    function to be dispatched to FaF
+///
+void Dispatch(Parallel::FireAndForget* faf, int32_t numEntries,
+              const std::function<void(int32_t)>& callback);
 
 }  // namespace Parallel
 }  // namespace PacBio