+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
  - Utility::Shuffle for portable alternative to std::shuffle
+ - Parallel::Dispatch to submit jobs to an existing FireAndForget
 
 ### Changed
  - Read.h pulse width is now optional to allow kinetic-free subreads as input